requirements.txt
tk
requests
pillow
yt-dlp